requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0

asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-env==1.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
